Move components out of Routes folder
