+++ app/routes/wow/shortages/single/index.tsx
@@ -5,14 +5,14 @@ import { InputWithLabel } from '~/components/form/InputWithLabel'
 import SmallFormContainer from '~/components/form/SmallFormContainer'
 import type { WowShortageResult } from '~/requests/WoWCommodities'
 import NoResults from '~/routes/queries/listings/NoResults'
-import WoWServerSelect from '../../full-scan/WoWServerSelect'
+import WoWServerSelect from '../../../../components/form/WoW/WoWServerSelect'
 import { validateShortageData, WoWShortageFormFields } from '../commodities'
 import ShortageResults from '../ShortageResults'
 import type { ActionFunction } from '@remix-run/cloudflare'
 import { json } from '@remix-run/cloudflare'
 import type { WOWSingleItemShortageProps } from '~/requests/WoWSingleItemShortage'
 import WoWSingleItemShortage from '~/requests/WoWSingleItemShortage'
-import { RegionRadioGroup } from '../../full-scan/RegionRadioGroup'
+import { RegionRadioGroup } from '../../../../components/form/WoW/RegionRadioGroup'
 import type { WoWServerRegion } from '~/requests/WOWScan'
 
 export const action: ActionFunction = async ({ request }) => {
